ezrax-agent/requirements.txt
scapy==2.5.0
paramiko==3.3.1
requests==2.31.0
psutil==5.9.6
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.1
jinja2==3.1.3
